Created files that share common functions
